+++ tests/testthat/test-tensor.R
@@ -89,7 +89,7 @@ test_that("Cuda tensor convertion", {
   skip_if_cuda_not_available()
 
   x <- torch_tensor(1, device = torch_device("cuda"))
-  expect_error(as_array(x), class = "runtime_error")
+  expect_error(as_array(x), regexp = NA)
 
   x <- x$to(dtype = torch_float(), device = torch_device("cpu"))
   expect_equal_to_r(x, 1)
@@ -586,4 +586,10 @@ test_that("grad_fn and cloning", {
   x <- torch_tensor(1, requires_grad = TRUE)
   x1 <- x$clone2()
   expect_true(grepl(pattern = "CloneBackward0", capture.output(x1$grad_fn), fixed = TRUE))
+})
+
+test_that("cuda tensor can be converted to tensor", {
+  skip_if_cuda_not_available()
+  x <- as.array(torch_tensor(1, device = "cuda"))
+  expect_equal(x, 1)
 })