+++ NEWS.md
@@ -8,6 +8,8 @@
 - Printing of `grad_fn` now appends a new line at the end.
 - Added support for a private `$finalize_deep_clone()` method for `nn_module` which
  allows to run some code after cloning a module.
+- Make sure deep cloning preserve state dict attributes. (#1129)
+- Converting torch tensor to R array works when tensor has 'cuda' device (#1130)
 
 # torch 0.12.0